extract errors pkg from service/storage
