+++ internal/errors/errors_test.go
@@ -1,4 +1,4 @@
-package storage
+package errors
 
 import (
 	"errors"
