+++ internal/service/storage/storage.go
@@ -5,25 +5,27 @@ import (
 	"io"
 )
 
+//
+
 type Storage interface {
 	// CreateObjectExclusively atomically creates an object named name with metadata metadata and data r if no object named name exists.
-	// Returns a non-nil error e such that ErrorIsCode(e, PreconditionFailed) is true if an object named name already exists.
-	//	 Implementors can create an error e such that ErrorIsCode(e, PreconditionFailed) is true by calling NewErrorf(PreconditionFailed, "...", ...).
+	// Returns an error e such that "github.com/go-mod-proxy/go-mod-proxy/internal/errors".ErrorIsCode(e, PreconditionFailed)
+	// is true if an object named name already exists.
 	CreateObjectExclusively(ctx context.Context, name string, metadata ObjectMetadata, data io.ReadSeeker) error
 
 	// DeleteObject deletes an object named name.
-	// Returns a non-nil error e such that ErrorIsCode(e, NotFound) is true if no object named name exists.
-	//	 Implementors can create an error e such that ErrorIsCode(e, NotFound) is true by calling NewErrorf(NotFound, "...", ...).
+	// Returns an error e such that "github.com/go-mod-proxy/go-mod-proxy/internal/errors".ErrorIsCode(e, NotFound)
+	// is true if no object named name exists.
 	DeleteObject(ctx context.Context, name string) error
 
 	// GetObject returns the data of an object as a reader.
-	// Returns a non-nil error e such that ErrorIsCode(e, NotFound) is true if no object named name exists.
-	//	 Implementors can create an error e such that ErrorIsCode(e, NotFound) is true by calling NewErrorf(NotFound, "...", ...).
+	// Returns an error e such that "github.com/go-mod-proxy/go-mod-proxy/internal/errors".ErrorIsCode(e, NotFound)
+	// is true if no object named name exists.
 	GetObject(ctx context.Context, name string) (io.ReadCloser, error)
 
 	// GetObjectMetadata returns the metadata of an object as a reader.
-	// Returns a non-nil error e such that ErrorIsCode(e, NotFound) is true if no object named name exists.
-	//	 Implementors can create an error e such that ErrorIsCode(e, NotFound) is true by calling NewErrorf(NotFound, "...", ...).
+	// Returns an error e such that "github.com/go-mod-proxy/go-mod-proxy/internal/errors".ErrorIsCode(e, NotFound)
+	// is true if no object named name exists.
 	GetObjectMetadata(ctx context.Context, name string) (ObjectMetadata, error)
 
 	// ListObjects returns a page of objects. A result *ObjectList o may have len(o.Names) < opts.MaxResults
